# Solver: Add individual flags to conflict sets.