--- a/cabal-install/Distribution/Solver/Types/Variable.hs
+++ b/cabal-install/Distribution/Solver/Types/Variable.hs
@@ -11,4 +11,4 @@ data Variable qpn =
     PackageVar qpn
   | FlagVar qpn FlagName
   | StanzaVar qpn OptionalStanza
-  deriving Eq
+  deriving (Eq, Show)