… cabal-install/Distribution/Solver/Modular/IndexConversion.hs
… cabal-install/Distribution/Solver/Modular/Linking.hs
… cabal-install/Distribution/Solver/Modular/Message.hs
… cabal-install/Distribution/Solver/Modular/Solver.hs
… cabal-install/Distribution/Solver/Modular/Tree.hs
… cabal-install/Distribution/Solver/Modular/Validate.hs
… cabal-install/Distribution/Solver/Modular/Var.hs
… cabal-install/cabal-install.cabal
… cabal-install/tests/UnitTests/Distribution/Solver/Modular/QuickCheck.hs
… cabal-install/tests/UnitTests/Distribution/Solver/Modular/Solver.hs