This PR implements #4391 and fixes #3930 and #4390.  It isn't ready to be merged yet, because I need to do some performance testing.  The change could have a big impact on solver run time for some packages, since it changes conflict counts.  Here are the main commits:

#### Solver: Add individual flags to conflict sets.

This commit changes the way that the solver tracks the variables that introduce
dependencies, in order to fix some bugs that prevented the solver from tracking
individual flags.  I refactored Dep, the type that represents a build-depends or
build-tool-depends dependency, so that it stores the package, flag, and stanza
choices that introduced the dependency.  That information is stored in a field
of type DependencyReason.  The DependencyReason is available to any solver phase
that needs to create conflict sets, such as "build" or "validation".  Whenever
there is a conflict involving a dependency, the solver creates a conflict set
and a log message using the dependency's DependencyReason.  This means that the
solver only needs to calculate the dependency reasons once, in
IndexConversion.hs, rather than every time they are used, i.e., in Builder.hs
(for goal reasons), Validate.hs, and Linking.hs.

Another difference between this commit and master is the dependencies between
solver variables.  On master, each dependency or variable is introduced by
exactly one other variable.  For example, if package x has a flag y, which
controls a dependency on package z, then the three variables depend on each
other in a chain, x -> y -> z.  If z can't be satisfied, the solver backjumps
to its cause, y, and if flipping y doesn't help, it continues backjumping to y's
cause, which is x.  In contrast, this commit allows each dependency to be
introduced by a package and any number of flags and stanzas.  So z's
DependencyReason would contain both x and y.

Storing all flags that introduced a dependency allows the solver to correctly
calculate conflict sets when there are nested conditionals.  We no longer need
to combine each package's flags into a single conflict set variable.  This
commit removes the `simplifyVar` function and adds flag variables directly to
conflict sets.  See issue #4391.

This commit makes another minor change.  In this commit and master, if a
dependency appears in the if and else blocks of a conditional, the solver lifts
it out of the conditional.  Previously, the solver behaved as if the flag did
not introduce the dependency.  This commit adds the flag variable to the
conflict set or error message if the dependency is involved in a conflict.  This
change doesn't affect correctness, but I think it improves the error messages,
since it gives the whole reason that each dependency was introduced.

#### Randomize the goal order in one of the dependency solver QuickCheck tests.

Completely randomizing the goal order exposes more bugs in backjumping than
using --reorder-goals.  I only applied the change to one test in this commit,
because the randomization function slowed down some of the other tests
significantly.

#### Add a test case for issue #4390.
#### Add a test case for a stanza preference bug (#3930).
#### Enable stanza preferences in the solver QuickCheck tests.

_____________________________________________________________________________

Here are some examples of changes in solver error messages that also reflect the changes in behavior.

`cabal install hackage-server`:

```diff
 Resolving dependencies...
 cabal: Could not resolve dependencies:
-trying: hackage-server-0.5.0:-build-hackage-import
+trying: hackage-server-0.5.0 (user goal)
 trying: hackage-server-0.5.0:+build-hackage-build
-trying: unix-2.7.2.0/installed-2.7... (dependency of
-hackage-server-0.5.0:+build-hackage-build)
-next goal: aeson (dependency of hackage-server-0.5.0:+build-hackage-build)
+trying: unix-2.7.2.0/installed-2.7... (dependency of hackage-server-0.5.0
++build-hackage-build)
+next goal: aeson (dependency of hackage-server-0.5.0 +build-hackage-build)
 rejecting: aeson-1.2.0.0, aeson-1.1.2.0, aeson-1.1.1.0, aeson-1.1.0.0,
 aeson-1.0.2.1, aeson-1.0.2.0, aeson-1.0.1.0, aeson-1.0.0.0, aeson-0.11.3.0,
 aeson-0.11.2.1, aeson-0.11.2.0, aeson-0.11.1.4, aeson-0.11.1.3,
 aeson-0.11.1.2, aeson-0.11.1.1, aeson-0.11.1.0, aeson-0.11.0.0, aeson-0.9.0.1,
 aeson-0.9.0.0, aeson-0.8.1.1, aeson-0.8.1.0, aeson-0.8.0.2, aeson-0.7.0.6,
-aeson-0.7.0.4, aeson-0.6.2.1, aeson-0.6.2.0 (conflict:
-hackage-server-0.5.0:build-hackage-build => aeson==0.6.1.*)
+aeson-0.7.0.4, aeson-0.6.2.1, aeson-0.6.2.0 (conflict: hackage-server
++build-hackage-build => aeson==0.6.1.*)
 rejecting: aeson-0.6.1.0 (conflict: unix => time==1.6.0.1/installed-1.6...,
 aeson => time<1.5)
 rejecting: aeson-0.6.0.2, aeson-0.6.0.1, aeson-0.6.0.0, aeson-0.5.0.0,
 aeson-0.4.0.1, aeson-0.4.0.0, aeson-0.3.2.14, aeson-0.3.2.13, aeson-0.3.2.12,
 aeson-0.3.2.11, aeson-0.3.2.10, aeson-0.3.2.9, aeson-0.3.2.8, aeson-0.3.2.7,
 aeson-0.3.2.6, aeson-0.3.2.5, aeson-0.3.2.4, aeson-0.3.2.3, aeson-0.3.2.2,
 aeson-0.3.2.1, aeson-0.3.2.0, aeson-0.3.1.1, aeson-0.3.1.0, aeson-0.3.0.0,
 aeson-0.2.0.0, aeson-0.1.0.0, aeson-0.10.0.0, aeson-0.8.0.1, aeson-0.8.0.0,
 aeson-0.7.0.5, aeson-0.7.0.3, aeson-0.7.0.2, aeson-0.7.0.1, aeson-0.7.0.0
-(conflict: hackage-server-0.5.0:build-hackage-build => aeson==0.6.1.*)
+(conflict: hackage-server +build-hackage-build => aeson==0.6.1.*)
 After searching the rest of the dependency tree exhaustively, these were the
-goals I've had most trouble fulfilling: aeson, hackage-server-0.5.0:flag,
-hackage-server-0.4:flag, template-haskell, hackage-server
+goals I've had most trouble fulfilling: hackage-server, aeson,
+hackage-server-0.5.0:build-hackage-build,
+hackage-server-0.4:build-hackage-mirror, template-haskell
```

There are several differences:

- The conflict set at the bottom contains the flag names, for example, `hackage-server-0.5.0:build-hackage-build` instead of `hackage-server-0.5.0:flag`.
- The summary of the steps that the solver took only lists the relevant flag choice, i.e., `hackage-server-0.5.0:+build-hackage-build`, instead of the two `hackage-server-0.5.0` flags that had been chosen at that point.
- The "conflict" message gives the value of the flag causing the conflict, for example, `(conflict: hackage-server +build-hackage-build => aeson==0.6.1.*)` instead of `(conflict: hackage-server-0.5.0:build-hackage-build => aeson==0.6.1.*)`.

`cabal install --dry-run git-annex -v3`:

```diff
 [__5] trying: ghc-prim-0.5.0.0/installed-0.5... (dependency of base-4.9.0.0/installed-4.9...)
 [__6] trying: unix-2.7.2.0/installed-2.7... (dependency of git-annex-6.20170519)
-[__7] trying: network-2.6.3.2 (dependency of git-annex-6.20170519)
+[__7] trying: network-2.6.3.2 (dependency of git-annex-6.20170519 +/-network-uri)
 [__8] trying: network-2.6.3.2:!test
 [__9] trying: split-0.2.3.2 (dependency of git-annex-6.20170519)
```

- This message shows that `network` is introduced by the `network-uri` flag, whether it is true or false.

/cc @kosmikus