@ezyang 2017-06-16
Let me know if you want an in depth code review, but the logic description in the commit sounds good to me.