@kosmikus 2017-06-16
Sounds fantastic. I'll try to take a look. But in any case, it's a good approach, and it should hopefully fix a few pathological cases we currently have. Looking forward to your performance test results.