@kosmikus 2017-06-21
@grayjay This all sounds rather good. For the cases where both master and branch hit the backjump limit, would it make sense to re-run with increased / infinite backjump limit?

Also, do you have any intuitive idea what's still difficult about these remaining cases?