@kosmikus 2017-06-21
@grayjay Also, for `nerf` with unlimited backjumps, how long does it take to find a solution?